+++ lang/cs/layouts.php
@@ -1,0 +1,14 @@
+<?php
+
+return [
+
+    'i18n.missing-all.description' => 'Chybí všechny překlady.',
+    'i18n.missing-all.title' => 'Testovací stránka pro překlady',
+    'i18n.missing-cs.description' => 'Chybí český překlad.',
+    'i18n.missing-cs.title' => 'Testovací stránka pro překlady',
+    'i18n.missing-en.description' => 'Chybí anglický překlad.',
+    'i18n.missing-en.title' => 'Testovací stránka pro překlady',
+    'i18n.complete.description' => 'Kompletní překlady.',
+    'i18n.complete.title' => 'Testovací stránka pro překlady',
+
+];