+++ resources/views/i18n/missing-en.blade.php
@@ -1,1 +1,1 @@
-<p>Missing en: {{ __('ig-common::messages.missing-en') }}</p>
+<p>{{ __('ig-common::messages.missing-en') }}</p>