+++ resources/views/i18n/test-trans.blade.php
@@ -1,3 +1,0 @@
-<p>Missing translation: {{ __('ig-common::messages.missing') }}</p>
-<p>Partially missing translation (fallback to en): {{ __('ig-common::messages.partially-missing') }}</p>
-<p>Translated: {{ __('ig-common::messages.translated') }}</p>