+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Add base view.
+- Add Translator test pages.
+
+### Fixed
+
+- Fix Translator missing translation detection and fallback.
+
 ## [0.5.4] - 2025-03-13
 
 ### Fixed
@@ -144,6 +155,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [0.5.4]: https://https://github.com/internetguru/laravel-common/compare/v0.5.3...v0.5.4
 [0.5.3]: https://https://github.com/internetguru/laravel-common/compare/v0.5.2...v0.5.3
 [0.5.2]: https://https://github.com/internetguru/laravel-common/compare/v0.5.1...v0.5.2